backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.10
psycopg2-binary==2.9.6
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.10.3
sqlalchemy-utils==0.41.1
fabric.js==0.7.0
python-dotenv==1.0.0
pytest==7.3.1
pytest-cov==4.0.0
